bnewton

+++ main.py
@@ -7,5 +7,5 @@ k=int(k)
 
 import math
 def newton(n, k): 
-    return math.factorial(n)/(math.factorial(k)*math.factorial(n-k))
+    return int(math.factorial(n)/(math.factorial(k)*math.factorial(n-k)))
 print(newton (n, k))
